% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_example.R
\name{download_example}
\alias{download_example}
\title{Download Example Files to Working Directory}
\usage{
download_example()
}
\description{
Copies the example files 'Form.xlsx' and 'Template.xlsx' from the
package's \code{extdata} folder to the user's current working directory.
}
\details{
The \code{download_example} function provides users with easy access to the
example datasets packaged with \code{reformR}. These datasets can
be used for demonstration purposes or to test the functionality of
other functions within the package.

The example files are stored in the \code{extdata} directory of the
\code{reformR} package. This function uses the \code{fs} package to handle
file operations.
}
\examples{
\dontrun{
# To copy the example files to your working directory
download_example()

# Check if the files are available in the working directory
list.files(pattern = "xlsx")
}

}
