% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{Transform Form Data Based on Template Metadata}
\usage{
template(df)
}
\arguments{
\item{df}{A data frame, typically a result from \code{read_excel()}, representing the raw Excel form.}
}
\value{
A modified tibble prepared for data extraction based on the template's metadata tags.
}
\description{
Uses a provided data frame, which is a read version of an Excel form, and transforms it
based on template metadata tags. The function prepares the data for further extraction
by renaming columns, adding index, pivoting, modifying, reclassifying, and filtering based
on specific conditions.
}
\details{
The \code{form_template} function is part of a larger process of automating data extraction from Excel forms.
These forms have been populated with specific tags in a template version to guide the data extraction.
Tags like "*((var." indicate single variables to be extracted. The function processes the raw data to only
keep the necessary meta-data tags and restructures the data in a usable format.

The process involves:
\enumerate{
\item Renaming columns to a consistent format using \code{rename_x()}.
\item Adding an index column to the data for tracking purposes using \code{add_index()}.
\item Pivoting the data to make it longer and more usable with \code{pivot_longer()}.
\item Modifying certain columns with \code{quickm()}.
\item Reclassifying columns to character data type with \code{reclass_c()}.
\item Filtering rows based on certain conditions, primarily to keep rows that have relevant tags using \code{filter_in()}.
}
}
\examples{
\dontrun{
library(readxl)

# Assume 'sample_form.xlsx' is a representative Excel form file.
raw_data <- read_excel("sample_form.xlsx")

transformed_data <- template(raw_data)
}

}
